requirements.txt
openai==0.27.2
python-dotenv==0.21.1
